+++ components/places/android/src/main/java/mozilla/appservices/places/RustError.kt
@@ -12,6 +12,7 @@ import com.sun.jna.Pointer
 import com.sun.jna.Structure
 import java.util.Arrays
 
+@Suppress("MagicNumber")
 internal open class RustError : Structure() {
 
     class ByReference : RustError(), Structure.ByReference
@@ -37,6 +38,7 @@ internal open class RustError : Structure() {
         return code != 0
     }
 
+    @Suppress("ComplexMethod", "ReturnCount", "TooGenericExceptionThrown")
     fun intoException(): PlacesException {
         if (!isFailure()) {
             // It's probably a bad idea to throw here! We're probably leaking something if this is