+++ gradle-plugin/src/main/kotlin/mozilla/appservices/MegazordDefinition.kt
@@ -28,7 +28,8 @@ data class MegazordDefinition(val name: String, val components: MutableSet<Modul
     private fun newId(identifier: String): ModuleIdentifier {
         val parts = identifier.split(':')
         if (parts.size != 2) {
-            throw GradleException("megazord moduleIdentifier must have 2 colon-separated parts; got: '$identifier' with ${parts.size} parts")
+            throw GradleException("megazord moduleIdentifier must have 2 colon-separated parts;" +
+                "got: '$identifier' with ${parts.size} parts")
         }
         return DefaultModuleIdentifier.newId(parts[0], parts[1])
     }