+++ CHANGELOG.md
@@ -8,6 +8,13 @@
 
 - Added a disableMemSecurity function to turn off some dubious behaviors of SQLcipher. ([#838](https://github.com/mozilla/application-services/pull/838))
 
+## Push
+
+### Breaking changes
+
+- `PushManager`'s `dispatch_for_chid` method has been renamed to `dispatchForChid`.
+- `PushManager` constructor arguments are now camelCased.
+
 # v0.21.0 (_2019-03-20_)
 
 [Full Changelog](https://github.com/mozilla/application-services/compare/v0.20.2...v0.21.0)